+++ src/pcap/tokio_impl/mod.rs
@@ -3,3 +3,7 @@
 //! Requires the `tokio-async` feature
 mod reader;
 pub use reader::AsyncPcapReader;
+
+pub mod buffer_pool;
+mod pooled_reader;
+pub use pooled_reader::AsyncPooledPcapReader;