+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `AsyncPcapReader::with_buf_reader` uses a predefined BufReader and returns `AsyncPcapReader<BufReader<R>>`
   - `AsyncPcapReader::new_without_buffer` is used when no buffer is desired and returns `AsyncPcapReader<R>`
 - Added Async Benchmarking
+- Added a pooled async pcap reader
 ## [0.5.0] (2026-01-30)
 - Added AnyPcapReader to read both pcap and pcapng files
 - Introduce AsyncPcapNgReader for asynchronous pcapng reading